fix(ci): preserve engine input extraction
