+++ lib/cluster-worker/engine-adapter.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { cloneJson } = require('./object-utils');
+const { inputFromRequest } = require('./engine-input');
 
 const TERMINAL_TOPICS = new Set(['CLUSTER_COMPLETE', 'CLUSTER_FAILED']);
 
@@ -25,17 +26,6 @@ function terminalEventFromMessage(message) {
     summary: data.reason || message.content?.text,
     code: data.code,
     reason: data.workerReason,
-  };
-}
-
-function inputFromRequest(request, artifactManifest) {
-  if (request.source === 'issue') return { issue: request.issue };
-  if (request.source === 'prompt') return { text: request.prompt };
-  return {
-    text:
-      'Execute the task described by this registry-staged, byte-free artifact manifest. ' +
-      'Do not request interactive input.\n' +
-      JSON.stringify(artifactManifest),
   };
 }
